==> experiments/02 Divide corpus into equal parts.ipynb <==
import os
import json
import random

from estnltk.converters import json_to_text

from modules.corpus_division import get_filenames,\
                                    get_hand_tagged, \
                                    get_labels_distribution, \
                                    get_documents_ne_statistics, \
                                    swap_items_in_lists, \
                                    calculate_proportions, \
                                    improve_scores, \
                                    remove_hand_tagged, \
                                    calculate_score, \
                                    n_even_chunks, \
                                    generate_random_division

# The location of protocols and the directories that have protocols in them
vallakohtufailid_location = os.path.join('..', 'data', 'vallakohtufailid')
directories = ["vallakohus_esimene", "vallakohus_teine", "vallakohus_kolmas", "vallakohus_neljas"]

# Location of the hand-tagged protocols file (contains id-s)
hand_tagged_protocols_file = os.path.join('..', 'data', 'k2sitsi_morfiga_protokollid.txt')

# Location of files that have the goldstandard tags (collected with the first tool)
json_files_location = os.path.join('..', 'data', 'vallakohtufailid-json-flattened')

# Location of the distributed corpus file
divided_corpus = os.path.join('..', 'data', 'divided_corpus.txt')

# Location of the hand-tagged files list
hand_tagged = os.path.join('..', 'data', 'corpus_subdistribution_without_hand_tagged.txt')

# Location of the list of files without goldstandard tags
no_goldstandard_tags_location = os.path.join('..', 'data', 'files_without_goldstandard_annotations.txt')

# Get all names of files in the corpus
filenames = get_filenames(vallakohtufailid_location, directories)

# Get labels and the number of that label from all documents
statistics = get_documents_ne_statistics(filenames, json_files_location)

# Get the distribution of labels over the whole corpus
ideal_distribution = get_labels_distribution(filenames, json_files_location)

# Get protocols that are tagged by hand
protocols_tagged_by_hand = get_hand_tagged(hand_tagged_protocols_file, filenames)

### Calculate the scores of subdistributions generated by the script

# Seed
seed = 1
random.seed(seed)
print(f'Seed: {seed}')

# Give a number of random divisions to be generated from the files
number_of_random_divisions = 25

# Give a number of how many even chunks to create from the files
number_of_even_chunks = 6

# Generate random distributions and divide them into subdistributions
random_distributions = generate_random_division(sorted(os.listdir(json_files_location)), number_of_random_divisions)
subdistributions = [n_even_chunks(distribution, number_of_even_chunks) for distribution in random_distributions]

sums_of_distributions = list()
list_of_scores = list()

for distribution in subdistributions:
    scores = dict()
    for files in distribution:
        score = calculate_score(files, statistics, ideal_distribution)
        scores[score] = files
        print(score)
    list_of_scores.append(scores)
    sub_corpus_score = sum(scores.keys()) / len(scores.keys())
    print(f'Sub-corpus score: {sub_corpus_score}\n')
    
sums_of_distributions = [(sum(k.keys()) / len(k.keys())) for k in list_of_scores]

# Find the smallest score in all of the randomly generated subdistributions
smallest_score = list_of_scores[sums_of_distributions.index(min(sums_of_distributions))]

# Get the smallest score files and find how many hand-tagged protocols they have
number_of_hand_tagged_protocols = [len(set(files).intersection(protocols_tagged_by_hand)) for files in smallest_score.values()]
files_to_use = [files for files in smallest_score.values()]
least_hand_tagged_old = files_to_use[number_of_hand_tagged_protocols.index(min(number_of_hand_tagged_protocols))]

# Remove files from the old list of least hand-tagged files
other_files = [files for files in files_to_use]
other_files.pop(other_files.index(least_hand_tagged_old))

# Remove hand-tagged files from least_hand_tagged by exchanging files from other subdistributions
least_hand_tagged, other_files = remove_hand_tagged(least_hand_tagged_old, other_files, statistics, ideal_distribution, protocols_tagged_by_hand)

# All files for the first five corpuses
final_files = [files for files in other_files]

# Write txt files
with open(divided_corpus, 'w+', encoding='UTF-8') as out_f:
    iterator = 1
    for files in final_files:
        for f in files:
            out_f.write(f'{str(f)}:{str(iterator)}\n')
        iterator += 1

with open(hand_tagged, 'w+', encoding = 'UTF-8') as out_f:
    for f in least_hand_tagged:
        out_f.write(f'{str(f)}:{str(iterator)}\n')

for files in final_files:
    print(f'Hand-tagged files in the subdistribution: {len(set(files).intersection(protocols_tagged_by_hand))}')
    print(f'Subdistribution score: {str(calculate_score(files, statistics, ideal_distribution))}\n')
print(f'Hand-tagged files in the subdistribution: {len(set(least_hand_tagged).intersection(protocols_tagged_by_hand))}')
print(f'Subdistribution score: {calculate_score(least_hand_tagged, statistics, ideal_distribution)}')

### Check which files don't have BRAT-annotated tags

no_goldstandard_annotations = list()
for file in filenames:
    with open(os.path.join(json_files_location, file), 'r', encoding='UTF-8') as in_f:
        text = json_to_text(in_f.read())
    if len(text.gold_ner) == 0:
        no_goldstandard_annotations.append(file)
        
with open(no_goldstandard_tags_location, 'w+', encoding='UTF-8') as out_f:
    for file in no_goldstandard_annotations:
        out_f.write(f'{file}\n')
