==> experiments/04 Display results of models.ipynb <==
import os
import json
import pandas as pd

from modules.results_extraction import results_by_subdistribution, results_by_named_entity, confusion_matrix

Define all the directories, which you want results for. These directories have to contain a `results.txt` file.

model_directories = [
    'model_default_with_vabamorftagger',
    'model_local_features_without_morph',
    'model_morph_without_lemmas',
    'model_morph_with_lemmas',
    'model_morph_with_lemmas_and_sentences',
    'model_morph_with_lemmas_and_sentences_and_gazzetteer',
    os.path.join('model_morph_with_lemmas_and_sentences_and_gazetteer_and_global_features', 'model_initial'),
    os.path.join('model_morph_with_lemmas_and_sentences_and_gazetteer_and_global_features', 'model_vabamorf_gazetteer'),
    os.path.join('model_morph_with_lemmas_and_sentences_and_gazetteer_and_global_features', 'model_vabamorf_gazetteer2'),
    os.path.join('model_morph_with_lemmas_and_sentences_and_gazetteer_and_global_features', 'model_vabamorf_gazetteer1and2'),
    os.path.join('model_morph_with_lemmas_and_sentences_and_gazetteer_and_global_features', 'model_gaz_loc'),
    os.path.join('model_morph_with_lemmas_and_sentences_and_gazetteer_and_global_features', 'model_gaz_loc_variants')    
]

Get files from the `divided_corpus.txt` file

divided_corpus = os.path.join('..', 'data', 'divided_corpus.txt')

files = {}

with open(divided_corpus, 'r', encoding = 'UTF-8') as in_f:
    txt = in_f.readlines()

for filename in txt:
    file, subdistribution = filename.split(":")
    files[file] = subdistribution.strip()

Get results from predefined model directories:

def get_results_from_directory(model_directory):
    with open(os.path.join('models', model_directory, 'results.txt')) as in_f:
        results_json = json.loads(in_f.read())
        return results_by_subdistribution(results_json, files)['Total'], results_by_named_entity(results_json, files)['Total']

totals_by_subdistribution = {}
totals_by_named_entity = {}

for model_directory in model_directories:
    by_subdistribution, by_named_entity = get_results_from_directory(model_directory)
    totals_by_subdistribution[model_directory] = by_subdistribution
    totals_by_named_entity[model_directory] = by_named_entity

# Total values by model:

display(pd.DataFrame(totals_by_subdistribution, index=["Precision", "Recall", "F1"]).T)

# Total values by named entity:

display(pd.DataFrame(totals_by_named_entity).T)

# Confusion matrices

files = []
with open(os.path.join('..', 'data', 'divided_corpus.txt'), 'r', encoding='UTF-8') as f:
    lines = f.readlines()

for line in lines:
    filename, subdistribution = line.strip().split(':')
    files.append(filename)

for model in model_directories:
    y_true, y_pred = confusion_matrix(model, files)
    print(model)
    print(pd.crosstab(y_true, y_pred))
    print('\n')
