==> experiments/06 Test today's NER tagger model on the court protocols.ipynb <==
# How well does today's Estonian written language NER model perform on the court protocols?

### Imports:

import json
import os
import re
import pandas as pd

from estnltk import Text
from estnltk.taggers import NerTagger
from estnltk.converters import text_to_json
from estnltk.converters import json_to_text
from estnltk.layer_operations import flatten

from nervaluate import Evaluator
from modules.preprocessing_protocols import preprocess_text
from modules.results_extraction import extract_results, results_by_subdistribution, results_by_named_entity

### Re-used variables:

divided_corpus = os.path.join('..', 'data', 'divided_corpus.txt')
json_files_location = os.path.join('..', 'data', 'vallakohtufailid-json-flattened')
baseline_files_location = os.path.join('..', 'data', 'vallakohtufailid_nertagger_baseline')
no_goldstandard_tags_location = os.path.join('..', 'data', 'files_without_goldstandard_annotations.txt')
results_location = os.path.join('..', 'data')

removed_layers = ['sentences', 'morph_analysis', 'compound_tokens', 'ner', 'words', 'tokens']
nertagger = NerTagger()

Read files from the distributed corpus:

files = {}

with open(divided_corpus, 'r', encoding = 'UTF-8') as f:
    txt = f.readlines()

for filename in txt:
    file, subdistribution = filename.split(':')
    files[file] = subdistribution.strip()

Get files without goldstandard annotations:

with open(no_goldstandard_tags_location, 'r', encoding='UTF-8') as in_f:
    lines = in_f.readlines()

no_goldstandard_annotations = [line.strip() for line in lines]

### Make `ner` and `wordner` layers from goldstandard files

for file in files:
    with open(os.path.join(json_files_location, file), 'r', encoding='UTF-8') as in_f:
        text = json_to_text(in_f.read()).text

        text = preprocess_text(Text(text))
        
        nertagger.tag(text)
        text.add_layer(flatten(text['ner'], 'flat_ner'))
                
        for x in removed_layers:
            text.pop_layer(x)
        
        text_to_json(text, file=os.path.join(baseline_files_location, file))

### Calculate the scores

extract_results(files,\
                no_goldstandard_annotations,\
                baseline_files_location,\
                json_files_location,\
                results_location)

### Results:

with open(os.path.join(results_location, 'results.txt'), 'r', encoding='UTF-8') as in_f:
    json_text = json.loads(in_f.read())
    results_subdist = results_by_subdistribution(json_text, files)
    results_ne = results_by_named_entity(json_text, files)

display(pd.DataFrame(results_subdist))

display(pd.DataFrame(results_ne))

Disclaimer: The results for the `LOC_ORG` and `MISC` tag are `0`, because today's written Estonian NER tagger model does not use these labels. `LOC_ORG` could either be defined as `LOC` or `ORG`.
